Refactored frontend tests
