+++ senz-web/frontend/src/__tests__/Notifier.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { shallow } from "enzyme";
-import Notifier from "./Notifier";
+import Notifier from "../components/Notifier";
 import { findByTestAttr, checkProps } from "../../Utils";
 
 describe("Notifier Component", () => {
