+++ senz-web/frontend/src/__tests__/Register.test.js
@@ -1,16 +1,16 @@
 import React from "react";
-import Login from "./Login";
-import { testStore, findByTestAttr } from "../../../Utils";
+import Register from "../components/authentication/Register";
+import { testStore, findByTestAttr } from "../../Utils";
 import { createShallow } from "@material-ui/core/test-utils";
 
-describe("Login Component", () => {
+describe("Register Component", () => {
   describe("Component Renders", () => {
     let shallow;
     let shallowComponent;
     beforeEach(() => {
       shallow = createShallow({ dive: true });
       const store = testStore({});
-      shallowComponent = shallow(<Login store={store} />)
+      shallowComponent = shallow(<Register store={store} />)
         .dive()
         .dive()
         .dive()
@@ -23,12 +23,12 @@ describe("Login Component", () => {
         .dive();
     });
     it("Should render the component correctly", () => {
-      const login = findByTestAttr(shallowComponent, "LoginComponent");
-      expect(login.length).toBe(1);
+      const register = findByTestAttr(shallowComponent, "RegisterComponent");
+      expect(register.length).toBe(1);
     });
     it("Should render the login form", () => {
-      const login = findByTestAttr(shallowComponent, "LoginForm");
-      expect(login.length).toBe(1);
+      const register = findByTestAttr(shallowComponent, "RegisterForm");
+      expect(register.length).toBe(1);
     });
   });
 });
